Extract KubernetesContextSelect to fix env-init dialog lint
EnvironmentDialogView.tsx tripped three lint errors on main after #386:
two no-unnecessary-condition hits on `dialog.error?.trim() ?? ''` (the
field is typed `string`, never null/undefined) and a max-lines:500
violation at 507 lines.

Move KubernetesContextSelect into its own file so the dialog drops
under 500, and replace the `?.`/`??` pair with a direct `.trim()`.
Behavior-preserving — same SelectField vs EmptyState rendering, same
Rescan handler, same "Last error from kubectl" suffix when
dialog.error is non-empty.

Closes #387
