+++ erun-ui/frontend/src/components/app/EnvironmentDialogView.tsx
@@ -1,7 +1,6 @@
 import { FolderPlus, LoaderCircle, Rocket } from 'lucide-react';
 import * as React from 'react';
 
-import { refreshKubernetesContexts } from '@/app/dialogContextsThunks';
 import {
   closeEnvironmentDialog,
   selectEnvironmentVersionSuggestion,
@@ -35,10 +34,9 @@ import { Label } from '@/components/ui/label';
 
 import { EditableComboField } from './EditableComboField';
 import { uniqueSuggestions } from './EditableComboField.helpers';
-import { EmptyState } from './EmptyState';
+import { KubernetesContextSelect } from './KubernetesContextSelect';
 import { ReadonlyField } from './ManageDialog.fields';
 import { RuntimeResourceControls } from './RuntimeResourceControls';
-import { SelectField } from './SelectField';
 import { VersionField } from './VersionField';
 
 const dialogErrorClassName =
@@ -267,55 +265,6 @@ function EnvironmentCreateFields({ dialog }: { dialog: EnvironmentDialog }): Rea
       />
       <EnvironmentCreateChecks dialog={dialog} />
     </>
-  );
-}
-
-function KubernetesContextSelect({ dialog }: { dialog: EnvironmentDialog }): React.ReactElement {
-  const dispatch = useAppDispatch();
-  const items = dialog.kubernetesContexts.map((context) => ({ value: context, label: context }));
-  const placeholder = dialog.kubernetesContextsLoading
-    ? 'Loading contexts...'
-    : 'Select Kubernetes context';
-  if (!dialog.kubernetesContextsLoading && dialog.kubernetesContexts.length === 0) {
-    const body =
-      "ERun runs `kubectl config get-contexts` using the PATH and KUBECONFIG it inherits from your login shell at startup. If your terminal sees contexts that don't appear here, set KUBECONFIG in ~/.zshenv (or ~/.bash_profile) so it applies to GUI launches too, then restart ERun. If kubectl is not yet installed, install it with `brew install kubectl`.";
-    const errorDetail = dialog.error?.trim() ?? '';
-    return (
-      <div className="grid gap-2">
-        <Label htmlFor="environment-kubernetes-context">Kubernetes context</Label>
-        <EmptyState
-          heading="No Kubernetes contexts found"
-          body={errorDetail !== '' ? `${body}\n\nLast error from kubectl:\n${errorDetail}` : body}
-          action={
-            <Button
-              type="button"
-              variant="outline"
-              size="sm"
-              onClick={() => {
-                void dispatch(refreshKubernetesContexts());
-              }}
-            >
-              Rescan
-            </Button>
-          }
-        />
-      </div>
-    );
-  }
-  return (
-    <SelectField
-      id="environment-kubernetes-context"
-      label="Kubernetes context"
-      value={dialog.kubernetesContext}
-      options={items}
-      placeholder={placeholder}
-      emptyLabel="No Kubernetes contexts"
-      disabled={dialog.busy || dialog.kubernetesContextsLoading}
-      required
-      onChange={(kubernetesContext) => {
-        dispatch(updateEnvironmentDialog({ kubernetesContext }));
-      }}
-    />
   );
 }
